Move all util methods to isContributionValid file, add const file for Resolvers.
