+++ backend/src/graphql/resolver/const/const.ts
@@ -1,0 +1,12 @@
+import Decimal from 'decimal.js-light'
+
+export const MAX_CREATION_AMOUNT = new Decimal(1000)
+export const FULL_CREATION_AVAILABLE = [
+  MAX_CREATION_AMOUNT,
+  MAX_CREATION_AMOUNT,
+  MAX_CREATION_AMOUNT,
+]
+export const CONTRIBUTIONLINK_NAME_MAX_CHARS = 100
+export const CONTRIBUTIONLINK_NAME_MIN_CHARS = 5
+export const CONTRIBUTIONLINK_MEMO_MAX_CHARS = 255
+export const CONTRIBUTIONLINK_MEMO_MIN_CHARS = 5
